+++ src/main/resources/Documentation/changelog.md
@@ -9,6 +9,11 @@ core
 
 bug fixed: When a tb reported `failed` status back -- discarding pending tasks -- the next `get` command still engaged with the purged tasks. The try to report with `put` for purged task failed with no such task error.
 
+test
+----
+
+complete job test and discarded tasks test added.
+
 Version 1.7: 2013-01-19
 ===========